requirements.txt
anyio==3.6.2
black==23.3.0
certifi==2023.5.7
click==8.1.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
imageio==2.31.1
imutils==0.5.4
lazy_loader==0.2
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
python-telegram-bot==20.3
PyWavelets==1.4.1
scikit-image==0.21.0
scipy==1.10.1
sniffio==1.3.0
tifffile==2023.4.12
